comfyui-yolo-api-node/requirements.txt
requests
Pillow
numpy
